\name{foodweb-package}
\alias{foodweb-package}
\alias{foodweb}
\docType{package}
\title{
  Food webs
}
\description{
  Generates food webs as ...
}

\details{
  \tabular{ll}{
    Package: \tab foodweb\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2011-09-01\cr
    License: \tab GNU Public License 2 or above\cr
  }
}

\author{ 
  Yangchen Lin  (Maintainer) <linyangchen@gmail.com>,

  Owen Petchey <owen.petchey@ieu.uzh.ch>,
  
  Karline Soetaert <k.soetaert@nioo.knaw.nl>,
  
  Bjšrn Rall <brall@gwdg.de>,
  
  Florian Schneider <f_schneider@bio.tu-darmstadt.de>
}

\references{
  
}

\examples{
\dontrun{
## show examples (see respective help pages for details)
example(generate.random)
example(Interval)
example(Benguela)

## open the directory with documents
browseURL(paste(system.file(package="foodwebs"), "/doc", sep=""))

## the vignette
%vignette("foodwebs")
}
}

\keyword{ package }