\name{moss}
\alias{moss}
\docType{data}
\title{Predator:prey matrix of a moss-microarthropod food web.}
\description{
Binary, symmetrical, ordered, predator:prey matrix. Column n represents the same species as row n. All species present in the food web are represented, including basal resources. Zeroes denote the absence of a trophic interaction. A one denotes that the species in the column consumes the species in the row.
}
\usage{moss}
\format{A data frame with 103 columns and rows.}
\source{Partial data for a publication that is in preparation.
Perdomo, G; Thompson, R; Sunnucks, P. (unpublished data).
}
\keyword{datasets}